Run workflow on events `push` and `pull_request`
My initial assumption was that `[push, pull_request]` is not a good
way to set up CI for public repositories.

In that scenario, only repo owners and collaborators push branches
on the original repo, everyone else is pushing to their own forks
and sends pull requests.

I also assumed the `pull_request` event is triggered more often
so public repo owner are forced to filter the activity type.

But it turns out Github provides sensible defaults, according to
https://help.github.com/en/actions/reference/events-that-trigger-workflows:

> Note: By default, a workflow only runs when a `pull_request`'s
> activity type is `opened`, `synchronize`, or `reopened`.
> To trigger workflows for more activity types, use the `types` keyword.

This means that `[push, pull_request]` is actually the most useful
setting: It runs when a pull request is opened or someone pushes more
commits to its base branch, and it runs on merges in the main repo.