Do not rely on ways tagged with postal_code. Fixes #27.
Previously, all (most?) ways we were looking for were tagged
with their respective postal_code.

Because of the bulk removal in
https://www.openstreetmap.org/changeset/87069135, the number of
primary roads with relevant primary ways dropped by 24%,
causing the build to fail.

This changes all queries to make use of the map_to_area feature
and the fact that in Germany, all postal codes are mapped as
relations.

What we do now is restrict to Berlin and search for those
postal codes relations. Then we map this to an area and
continue searching for relevant ways.

This takes slightly longer but produces good results.