Trick Make into thinking everything is up-to-date
Because Make checks if prerequisites are newer than e.g.
dist/berlin/roads.json, it wants to remake the whole chain
of targets.

This is not what we want. Instead, by touching all files
instead of running their commands, we pretend the commands
were done.

Before, we have to create the tmp directory for touch
to succeed.